recovery: change delivery_rate to u64
delivery_rate is defined as bytes/sec, so no need to be f64.

+++ src/lib.rs
@@ -3163,7 +3163,7 @@ pub struct Stats {
     pub cwnd: usize,
 
     /// The estimated data delivery rate in bytes/s.
-    pub delivery_rate: f64,
+    pub delivery_rate: u64,
 }
 
 impl std::fmt::Debug for Stats {
